# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AlustarsCRM API",
    "version": "1.0.0",
    "description": "AlustarsCRM 客户关系管理系统 API 文档"
  },
  "servers": [
    {
      "url": "http://localhost:3333",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "object": {},
      "OrderItemSchema": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "product": {
            "type": "string"
          },
          "product_name": {
            "type": "string"
          },
          "product_code": {
            "type": "string"
          },
          "part_number": {
            "type": "string"
          },
          "description_en": {
            "type": "string"
          },
          "quantity": {
            "type": "number",
            "minimum": 0,
            "exclusiveMinimum": true
          },
          "unit": {
            "type": "string"
          },
          "unit_price": {
            "type": "number",
            "minimum": 0
          },
          "cost_price": {
            "type": "number",
            "minimum": 0
          }
        },
        "required": [
          "product",
          "product_name",
          "product_code",
          "quantity",
          "unit",
          "unit_price"
        ]
      },
      "OrderSchema": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "customer_id": {
            "type": "string"
          },
          "project_id": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "total_amount": {
            "type": "number"
          },
          "currency": {
            "type": "string"
          },
          "payment_terms": {
            "type": "string"
          },
          "delivery_date": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "code",
          "customer_id",
          "status",
          "total_amount",
          "currency",
          "created_at",
          "updated_at"
        ]
      },
...
      "PurchaseOrderPayment": {},
...
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ]
      }
    },
    "responses": {
      "400": {
        "description": "Bad Request",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Invalid request parameters"
                }
              }
            }
          }
        }
      },
...
      "500": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "An unexpected error occurred"
                }
              }
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/disk/list": {
      "get": {
        "operationId": "get-disk-list",
        "summary": "列出 S3 存储目录内容",
        "description": "根据前缀列出文件和文件夹，支持递归模式获取所有文件",
        "tags": [
          "Disk"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "返回 { folders:FolderInfoSchema[], files:FileInfoSchema[] } 格式的目录内容",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/object"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/disk/upload": {
      "post": {
        "operationId": "post-disk-upload",
        "summary": "上传文件到 S3存储",
        "description": "支持直接上传文件到S3，可指定文件夹或完整路径。文件大小限制为100MB。",
        "tags": [
          "Disk"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "上传成功，返回文件路径、名称、大小等信息",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
    "/orders/from-template": {
      "post": {
        "operationId": "post-orders-from-template",
        "summary": "Create an order from a template",
        "description": "Creates a new sales order by copying data from a saved template. Generates a compact order code, applies customer/project overrides, and copies template items into the new order.",
        "tags": [
          "Orders"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Order created from template",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderSchema"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
    "/orders/templates": {
      "get": {
        "operationId": "get-orders-templates",
        "summary": "List all order templates",
        "description": "Retrieves all saved order templates, sorted by creation date descending.",
        "tags": [
          "Orders"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Order templates list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/200 templates"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "post": {
        "operationId": "post-orders-templates",
        "summary": "Create an order template from an existing order",
        "description": "Creates a reusable template by copying data and items from an existing sales order. Requires a name and source order ID.",
        "tags": [
          "Orders"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Template created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/200 template"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
    "/orders/{id}/items-with-shipped": {
      "get": {
        "operationId": "get-orders-{id}-items-with-shipped",
        "summary": "获取订单项及已发货数量",
        "description": "获取指定订单的所有明细项，并计算每项的已发货数量和剩余数量",
        "tags": [
          "Orders"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number"
            },
            "example": 123,
            "description": "Path parameter: id"
          }
        ],
        "responses": {
          "200": {
            "description": "订单项列表（含已发货数量）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderItemSchema"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
    "/purchase-orders/{id}/payments": {
      "get": {
        "operationId": "get-purchase-orders-{id}-payments",
        "summary": "",
        "description": "",
        "tags": [
          "Purchase-orders"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number"
            },
            "example": 123,
            "description": "Path parameter: id"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PurchaseOrderPayment"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "post": {
        "operationId": "post-purchase-orders-{id}-payments",
        "summary": "",
        "description": "",
        "tags": [
          "Purchase-orders"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number"
            },
            "example": 123,
            "description": "Path parameter: id"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurchaseOrderPayment"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
  },
...
}